Move Ethereum network to http_api module
Now we only use ChainId internally, the network is only used for
backward compatibility purposes on the http api interface.
